Issue para Reestructuración del proyecto para implementación del CLI
