+++ src/cli.py
@@ -1,0 +1,7 @@
+"""
+Computación I - Sistema de Gestión para una Clínica.
+Nombre y Apellido: Enzo Agustín Aguirre Polenta.
+Ciclo Lectivo: 2025.
+Carrera: Ingeniería en Informática.
+Ruta: "computaci-n-2025-05-27-gesti-n-cl-nica-EnzoAguirre04/src/cli.py".
+"""
