+++ src/modelo/__init__.py
@@ -3,5 +3,5 @@ Computación I - Sistema de Gestión para una Clínica.
 Nombre y Apellido: Enzo Agustín Aguirre Polenta.
 Ciclo Lectivo: 2025.
 Carrera: Ingeniería en Informática.
-Ruta: "/tests/test_main.py".
+Ruta: "computaci-n-2025-05-27-gesti-n-cl-nica-EnzoAguirre04/src/modelo/__init__.py".
 """
